optimize: swow.php
