--- examples/web-skeleton/src/Container/Swow/Exception/CoroutineDestroyedException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Container\Swow\Exception;
-
-class CoroutineDestroyedException extends \RuntimeException
-{
-
-}
